Extract move-native vector routines to module
